Adds 'flow'
Goal is for 'flow' to supplement or replace 'tflow' and 'clflow'
but the use remains fairly involved, requiring setting each option
manually. Great for invocation but hard for setup.

Its general use is more suitable for non-tutorial purposes
but as the "undocumented" options show, it is still tuned for
learning use, especially in-house learning.